+++ Packages/com.nonatomic.assembly-template/Editor/Source/AssemblyAttrs.cs
@@ -1,0 +1,1 @@
+using System.Runtime.CompilerServices;[assembly: InternalsVisibleTo("AssemblyTemplate.Tests.PlayMode")][assembly: InternalsVisibleTo("AssemblyTemplate.Tests.EditMode")]