+++ Packages/com.nonatomic.assembly-template/Tests/EditMode/Thing.cs
@@ -1,0 +1,7 @@
+﻿namespace AssemblyTemplate.Tests.EditMode
+{
+	public class Thing
+	{
+		
+	}
+}